exam_backend/requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
diff-match-patch==20241021
Django==5.2.7
django-cors-headers==4.4.0
django-filter==25.1
django-health-check==3.20.0
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-standardized-errors==0.14.1
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.3.0
kombu==5.5.4
msgpack==1.1.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
pytest==8.4.2
pytest-django==4.11.1
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
openpyxl==3.1.5
xlrd==2.0.1
zope.interface==7.2
